requirements.txt
boto3==1.16.41
botocore==1.19.41
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
numpy==1.19.4
opencv-python==4.4.0.46
Pillow==8.0.1
python-dateutil==2.8.1
s3transfer==0.3.3
six==1.15.0
urllib3==1.26.2
Werkzeug==1.0.1
